refactor: extract drag gesture logic into a reusable modifier
